pyproject.toml
[tool.poetry]
name = "web-hw-13"
version = "0.1.0"
description = ""
authors = ["AlexanderF048 <alexandr.frankevich@gmail.com>"]
readme = "README.md"
packages = [{include = "web_hw_13"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.20.0"
fastapi = "^0.93.0"
sqlalchemy = "^2.0.5.post1"
libgravatar = "^1.0.4"
fastapi-mail = "^1.2.6"
alembic = "^1.10.1"
psycopg2 = "^2.9.5"
pydantic = {extras = ["email"], version = "^1.10.5"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.6"
redis-lru = "^0.1.2"
python-dotenv = "^1.0.0"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.32.0"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
